% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simode_profile.R
\name{confint.profile.simode}
\alias{confint.profile.simode}
\title{Calculates confidence intervals for the model parameters}
\usage{
\method{confint}{profile.simode}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted model object: \code{profile.simode} object returned by a call to \code{\link{profile}}.}

\item{parm}{A specification of which parameters are to be given confidence intervals (named vector).
If missing, all parameters are considered.}

\item{level}{The confidence level required.}

\item{...}{Additional argument(s) for methods.}
}
\value{
The confidence intervals.
}
\description{
Calculates confidence intervals for the model parameters,
based on the given likelihood profiles
}
