% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simode_profile.R
\name{plot.confint.simode}
\alias{plot.confint.simode}
\title{Plot confidence intervals for the model parameters}
\usage{
\method{plot}{confint.simode}(x, which = NULL, pars_true = NULL,
  legend = F, cols = list(fit = "blue", true = "black"), ...)
}
\arguments{
\item{x}{\code{confint.simode} object returned by a call to \code{\link{confint}}}

\item{which}{Which parameters to plot the confidence intervals for.
If empty, the plot will include all of the parameters in x.}

\item{pars_true}{The true parameter values (if are known).}

\item{legend}{Whether or not to add a figure legend.}

\item{cols}{List of colors for each element of the plot.}

\item{...}{Additional argument(s) for methods.}
}
\description{
Plot confidence intervals for the model parameters of a \code{simode} object,
calculated based on likelihood profiles
}
