% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simode_profile.R
\name{plot.profile.simode}
\alias{plot.profile.simode}
\title{Plot the likelihood profiles for the model parameters}
\usage{
\method{plot}{profile.simode}(x, which = NULL, mfrow = par("mfrow"),
  cols = list(fit = "blue", threshold = "red"), ...)
}
\arguments{
\item{x}{\code{profile.simode} object returned by a call to \code{\link{profile}}.}

\item{which}{Which parameters to plot the likelihood profiles for.
If empty, the plot will include all of the parameters in x.}

\item{mfrow}{A vector of the form c(nr,nc) setting the layout of
subplots in one plot (see also \code{\link{par}}).}

\item{cols}{List of colors for each element of the plot.}

\item{...}{Additional argument(s) for methods.}
}
\description{
Plot the likelihood profiles for the model parameters
}
