% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simode_profile.R
\name{print.confint.simode}
\alias{print.confint.simode}
\title{Print method for \code{confint.simode} objects}
\usage{
\method{print}{confint.simode}(x, ...)
}
\arguments{
\item{x}{The \code{confint.simode} object.}

\item{...}{Additional argument(s) for methods.}
}
\description{
Print method for \code{confint.simode} objects
}
