% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simode_profile.R
\name{print.profile.simode}
\alias{print.profile.simode}
\title{Print method for \code{profile.simode} objects}
\usage{
\method{print}{profile.simode}(x, ...)
}
\arguments{
\item{x}{The \code{profile.simode} object.}

\item{...}{Additional argument(s)for methods.}
}
\description{
Print method for \code{profile.simode} objects
}
