% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simode.R
\name{print.summary.simode}
\alias{print.summary.simode}
\title{Print method for \code{summary.simode} objects}
\usage{
\method{print}{summary.simode}(x, ...)
}
\arguments{
\item{x}{The \code{summary.simode} object.}

\item{...}{Additional argument(s) for methods.}
}
\description{
Print method for \code{summary.simode} objects
}
