% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{solve_ode2}
\alias{solve_ode2}
\title{Ordinary differential equations solver using a \code{simode} object}
\usage{
solve_ode2(x, type = c("both", "im", "nls"))
}
\arguments{
\item{x}{A simode object returned from a call to \code{\link{simode}}.}

\item{type}{Which solution to generate ('both'\\'nls'\\'im').}
}
\value{
A matrix whose first column contains the given time points
and subsequent columns hold the computed ODE equations' values at these time points.
}
\description{
A wrapper for the \link[deSolve]{ode} function that solves a system of
ordinary differential equations described using symbolic equations.
}
