% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simode.R
\name{summary.simode}
\alias{summary.simode}
\title{Summary method for \code{simode} objects}
\usage{
\method{summary}{simode}(object, digits = max(3, getOption("digits") -
  3), ...)
}
\arguments{
\item{object}{The \code{simode} object.}

\item{digits}{The number of significant digits to use.}

\item{...}{Additional argument(s) for methods.}
}
\description{
Summary method for \code{simode} objects
}
